backend/requirements.txt
Flask==3.0.0
mysql-connector-python==8.3.0
flask-bcrypt==1.0.1
PyJWT==2.8.0
flask-cors==4.0.0
requests==2.31.0
geopy==2.4.1
